+++ src/main/java/com/app/jungsuri/domain/mountain/persistence/MountainLocationEntity.java
@@ -1,0 +1,2 @@
+package com.app.jungsuri.domain.mountain.persistence;public class MountainLocationEntity {
+}
