+++ src/main/java/com/app/jungsuri/domain/mountain/persistence/MountainLocationService.java
@@ -1,0 +1,2 @@
+package com.app.jungsuri.domain.mountain.persistence;public class MountainLocationService {
+}
